Improve CLI args, bump Scala version
The CLI args were inconsistent, this unifies on an approach in the
docs, tests and implementation.

We were on Scala 3.4.x to benefit from the for-comprehension
improvements in that release. That version has been supplanted by
Scala 3.6.x so we update to that version. Unfortunately that brings
some collateral changes (we have to specify the `-source` option, and
the new context bounds syntax - which also requires updating the
scalafmt configuration).