Add support for markdown file output
To help with GitHub actions integration, the output can now be a
Markdown formatted file, instead of the terminal. This requires a
different output format, since terminal control codes won't work in
the file.

To make this work, we add a "formatter" abstraction, and ask the
"printer" to specify its format.